+++ tailwind.config.js
@@ -1,6 +1,6 @@
 /** @type {import('tailwindcss').Config} */
 export default {
-  content: ['./src/client/index.html', './src/client/**/*.{ts,tsx}'],
+  content: ['./app/client/index.html', './app/client/**/*.{ts,tsx}'],
   theme: {
     extend: {
       colors: {